+++ cli/cli/Commands/Project/NewStorageCommand.cs
@@ -30,7 +30,7 @@ public class NewStorageCommand : AppCommand<NewStorageCommandArgs>, IStandaloneC
 				() =>
 				{
 					var currentVersion = AppContext.TargetFrameworkName.Split('=')[1].Substring(1);
-					var currentVersionDouble = double.Parse(currentVersion);
+					var currentVersionDouble = double.Parse(currentVersion, System.Globalization.CultureInfo.InvariantCulture);
 					if (currentVersionDouble < 8.0)
 					{
 						currentVersion = "8.0";